Change taps type based on mode
